+++ packages/libjpeg-turbo-12bit/CHANGELOG.md
@@ -3,6 +3,13 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.4.6](https://github.com/cornerstonejs/codecs/compare/@cornerstonejs/codec-libjpeg-turbo-12bit@0.4.5...@cornerstonejs/codec-libjpeg-turbo-12bit@0.4.6) (2026-09-09)
+
+**Note:** Version bump only for package @cornerstonejs/codec-libjpeg-turbo-12bit
+
+
+
+
 ## [0.4.5](https://github.com/cornerstonejs/codecs/compare/@cornerstonejs/codec-libjpeg-turbo-12bit@0.4.4...@cornerstonejs/codec-libjpeg-turbo-12bit@0.4.5) (2026-09-02)