+++ packages/libjpeg-turbo-8bit/CHANGELOG.md
@@ -3,6 +3,13 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.2.7](https://github.com/cornerstonejs/codecs/compare/@cornerstonejs/codec-libjpeg-turbo-8bit@1.2.6...@cornerstonejs/codec-libjpeg-turbo-8bit@1.2.7) (2026-09-09)
+
+**Note:** Version bump only for package @cornerstonejs/codec-libjpeg-turbo-8bit
+
+
+
+
 ## [1.2.6](https://github.com/cornerstonejs/codecs/compare/@cornerstonejs/codec-libjpeg-turbo-8bit@1.2.5...@cornerstonejs/codec-libjpeg-turbo-8bit@1.2.6) (2026-09-02)